[CD-210] Rename: tests to utils
